fix: address review comment 4589077469 — reorganize gc worktree tests
